+++ .gitignore
@@ -17,3 +17,5 @@ venv/
 
 dist/
 docs/_build
+
+.DS_Store
