+++ docs/backends/gcloud.rst
@@ -86,7 +86,7 @@ inferred from the environment.
 ``GS_IS_GZIPPED`` (optional: default is ``False``)
     Whether or not to enable gzipping of content types specified by ``GZIP_CONTENT_TYPES``
 
-``GZIP_CONTENT_TYPES`` (optional: default is ``text/css``, ``text/javascript``, ``application/javascript``, ``application/x-javascript``, ``image/svg+xml``, ``application/wasm``)
+``GZIP_CONTENT_TYPES`` (optional: default is ``text/css``, ``text/javascript``, ``application/javascript``, ``application/x-javascript``, ``image/svg+xml``)
     When ``GS_IS_GZIPPED`` is set to ``True`` the content types which will be gzipped
 
 .. _gs-creds:
