+++ setup.cfg
@@ -54,7 +54,8 @@ sftp =
 [flake8]
 exclude =
     .tox,
-    docs
+    docs,
+    venv
 max-line-length = 119
 
 [isort]
